requirements.txt
attrs==24.2.0
Automat==24.8.1
blinker==1.8.2
buildtools==1.0.6
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.3
docopt==0.6.2
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
furl==2.1.3
gcloud==0.18.3
googleapis-common-protos==1.65.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
MarkupSafe==3.0.2
oauth2client==4.1.3
orderedmultidict==1.0.1
protobuf==5.29.0rc2
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pyparsing==3.2.0
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==4.1.0
redo==3.0.0
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
setuptools==75.5.0
simplejson==3.19.3
six==1.16.0
SQLAlchemy==2.0.36
Twisted==24.10.0
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==3.1.2
wheel==0.45.0
zope.interface==7.1.1
